Library

Reduce.jl Library

Reduce.jl Library

Functions

Index

Reduce.jl

Introduction

Setup

Background

Usage

Troubleshooting

Functions

Reduce.Reset — Function.

Reduce.Reset()

Kills the REDUCE process and starts a new instance.

Examples

julia> Reduce.Reset()
Reduce (Free PSL version, revision 4015),  5-May-2017 ...

source

Reduce.RExpr — Type.

A Reduce expression

Summary:

type RExpr <: Any

Fields:

str :: Array{Compat.String,1}

source

Reduce.rcall — Function.

rcall(r::RExpr)

Evaluate a Reduce expression.

Examples

julia> R"int(sin(x), x)" |> RExpr |> rcall
 - cos(x)

source

rcall{T}(e::T)

Evaluate a Julia expression or string using the Reduce interpretor and convert output back into the input type

Examples

julia> rcall("int(sin(y)^2, y)")
"( - cos(y)*sin(y) + y)/2"
julia> rcall(:(int(1/(1+x^2), x)))
:(atan(x))

source

Base.parse — Function.

parse(r::RExpr)

Parse a Reduce expression into a Julia expression

Examples

julia> parse(R"sin(i*x)")
:(sinh(x) * im)

source

parsegen

bematch

Index

Reduce.RExpr

Base.parse

Reduce.Reset

Reduce.rcall